moved http abstraction since we will reuse
